#1404 Add support for cotable id in EnumerateIntents
Summary:
YSQL system catalog data is stored in the master Raft group. A "cotable id" is a UUID that is stored
as a prefix of all DocDB keys of a particular YSQL system table, to allow multiple such tables to
exist in the same RocksDB.  To support transactional YSQL DDL, we need to be able to correctly
generate the set of weak intents for a DocKey that includes a cotable id.

For example, if a SubDocKey is (cotable_id1, hash, (h1, h2), (r1, r2), (subkey1)), then the set of
weak intents that we generate is now as follows:

(cotable_id1)
(cotable_id1, hash, (h1, h2))
(cotable_id1, hash, (h1, h2), (r1))
(cotable_id1, hash, (h1, h2), (r1, r2))

Note that we are not generating an empty weak intent for a key with a cotable id anymore, because we
don't want to lock all YSQL system tables with such an intent, only a particular system table.

Test Plan: Jenkins

Reviewers: sergei, mihnea

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D7178