Run YSQL system catalog updates that are part of DDL operations in separate DocDB transactions
Summary:
This diff makes a step towards enabling transactional DDL by switching to transactional execution of multiple YSQL system catalog read/write operations that happen as part of the same DDL operation.

We make YSQL system tables transactional, backed by the regular transaction status table that resides on tablet servers. This introduces an additional dependency of the master on tablet servers, so that accessing YSQL system catalog tables now in the common case requires tablet servers to be up and running. This could be addressed later by making the system catalog tablet a transaction status tablet in addition to everything else it already stores.

After transactional capabilities are added to YSQL system tables, there are multiple possible choices in how to wrap formerly non-transactional DDL operations that access the YSQL system catalog into transactions. Ideally, we would want to allow arbitrary user-controlled transactions to include DDL statements that could be committed/rolled back similarly to DML statements. That would require us to commit and rollback the corresponding Yugabyte-level operations (creating namespaces, tablets, and tablets in the Yugabyte system catalog, which is also stored in the master but is distinct from the YSQL system catalog) along with YSQL system catalog changes. However, this diff's goal is more modest: it is an incremental improvement from the current state of fully non-transactional YSQL system catalog changes that would allow a reader to e.g. see a state of a half-done CREATE TABLE statement. With this diff, all DDL statements from a small list (currently just CREATE TABLE) are wrapped in a separate DocDB-level transaction that is outside of the regular transaction that DML statements participate in.

Test Plan: Jenkins

Reviewers: mihnea, neha, sergei

Subscribers: bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D7156