Run tests on Spark without NFS
Summary:
Our old approach for running tests on Spark was to access any files
needed to run the test over NFS exported by the Jenkins slave. This
caused unpredictable test errors and lock-ups.

Here we implement the following approach:
- Create a tar.gz with all code and compilation output needed to run the
  tests remotely.
- Distribute the tar.gz to all workers using PySpark's addFile facility.
- Extract the archive to the exact same path where it was on the main
  build host (e.g. Jenkins slave). On the worker, this path is not an NFS
  path but a local path. Permissions must be set correctly on the parent
  directory of all test workspace directory so we can do that.
- Run the tests on the build worker.
- Copy any relevant test output files (test log, JUnit-compatible XML
  file, our own JSON-based test report, etc.) back to the main build
  worker using a separate NFS mount. The Spark worker only uses NFS at
  this step.

Also simplifying the process of finding the Linuxbrew directory. Creating
a file called `linuxbrew_path.txt` in the build directory, and if that
file is present, always using the Linuxbrew directory from that file
instead of trying to auto-detect it. This requires determining the build
directory before we can determine the Linuxbrew path, and changing the
auto-detection of whether to use Ninja instead of Make.

Also removing Python interpreter wrapper scripts. We should just install
the relevant version of Python on the build system.

Here is how we collect all the test file patterns that a particular test
produces, so that we can copy them to the main build node. Adding a
function register_test_artifact_files in common-test-env.sh and calling
it whenever we determine a location for log files. Sometimes the
arguments are patterns ending with a `*`. Then, in the worker-side
function in run_tests_on_spark.py, we copy files matching these patterns
back to the main build node.

Test Plan: Jenkins: build platform: mac

Reviewers: wesley, bogdan

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D7801