Summary: Our old approach for running tests on Spark has been such that we would access any files needed to run the test over NFS.  This might have been causing unpredictable test errors and NFS-related lock-ups.
In this diff we implement the following alternative approach for running
tests on Spark:
- Create a tar.gz with all code and compilation output needed to run the
  tests remotely.
- Distribute the tar.gz to all workers using PySpark's addFile facility.
- Extract the archive to the exact same path where it was on the main
  build host (e.g. Jenkins slave). On the worker, this path is not an NFS
  path but a local path. Permissions must be set correctly on the build
  worker node for this to work.
- Run the tests on the build worker.
- Copy any relevant test output files (test log, JUnit-compatible test
  result XML files, our own JSON-based test reports, etc.) back to the
  main build node using a separate NFS mount. This is the only step at
  which we use NFS in the new workflow, and this could be later replaced
  with e.g. scp or sshfs.

Details of collecting test result paths in a Spark task
-------------------------------------------------------

Here is how we collect all the test file patterns that a particular test
produces, so that we can copy them to the main build node. Adding a
function register_test_artifact_files in common-test-env.sh and calling
it whenever we determine a location for log files. Sometimes the
arguments are patterns ending with a `*`. Then, in the worker-side
function in run_tests_on_spark.py, we copy files matching these patterns
back to the main build node.

Persisting paths of third-party dependencies and Linuxbrew/Homebrew
-------------------------------------------------------------------

Also simplifying the process of finding the Linuxbrew directory. We are
now creating a file called `linuxbrew_path.txt` in the build directory,
and if that file is present, we will always use the Linuxbrew directory
from that file instead of trying to auto-detect it. This requires
determining the build directory before we can determine the Linuxbrew
path, and changing the mechanism of auto-detection of whether to use
Ninja instead of Make. (Availability of Ninja could previously influence
the build directory path). Also creating similar files for third-party
dependencies' path and URL (`thirdparty_path.txt` and
`thirdparty_url.txt') and custom Homebrew directory on macOS
(`custom_homebrew_path.txt`).

Removing Python wrapper scripts
-------------------------------

Removing Python interpreter wrapper scripts and not using Python fron
Homebrew/Linuxbrew anymore. The build system will now use the default
installation of python2.7.

Download third-party dependencies into a fixed local path on Linux
------------------------------------------------------------------

On Linux (CentOS) we are enabling the `--download-thirdparty` mode by
default in Jenkins. This mode, introduced in de42d48142, enables
automatic installation of third-party dependencies and Linuxbrew from
URLs on GitHub into fixed directories under `/opt/yb-build`. This
improves compilation speed and also eliminates another path of NFS
access on Spark workers when running tests. We will also enable this on
macOS when we build the necessary archives.

The yb_release.py script needed to be updated to be able to pick up the
correct Linuxbrew location from the `linuxbrew_path.txt` file in the
build directory and correctly classify executables and dynamic libraries
residing under that path.

Local Maven repository changes
------------------------------
For Linux Jenkins jobs we now put the Maven local repository into the
`build/m2_repository` directory (where `build` is the parent directory
of build directories for various build types). We use Maven Dependency
Plugin's goals such as `dependency:resolve` and
`dependency:go-offline` to force Maven to download the most complete set
of artifacts needed by the project. Then we package this local
repository as part of the archive shipped to Spark workers, so that no
Maven artifact downloads have to happen as part of Spark tasks that run
YugabyteDB Java tests.

Other changes
-------------

- Reducing verbosity of logging during C++ and Java compilation and in
  CMake.
- Use addPyFile with a zip archive containing the contents of our
  `python` directory to ship the Python code of the `yb_dist_tests.py`
  module and its dependencies to Spark workers instead of ad-hoc
  sys.path manipulations.
- Added a `--no_reinitdb` flag to the top-level yb_release script to
  avoid re-running initdb every time. This helps with debugging the
  release script.

Test Plan:
Jenkins

To locally debug running tests on Spark use e.g.:

```
  spark-submit build-support/run_tests_on_spark.py \
    --build-root build/release-gcc-dynamic-ninja \
    --cpp \
    --cpp_test_program_regexp=ql-dml-test \
    --spark-master-url=<spark_url> \
    --send_archive_to_workers
```

Reviewers: wesley, bogdan

Reviewed By: bogdan

Subscribers: jenkins-bot, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D7801