[#4411] Add a gflag to log all operations on RocksDB iterators
Summary:
Add a gflag to log all operations on RocksDB iterators. This is really
useful for understanding the overhead of the read path in different
scenarios. Currently there is a gflag that enables logging of all
RocksDB operations on newly created iterators to the regular glog
stream, and it is suitable for use in pg_mini-test. The logging wrapper
is at RocksDB level and does no DocDB aware key/value decoding yet.

Test Plan:
Jenkins

New tests (inserting a few rows and overwriting them many times, and
then logging the scan operations). The tests are not checking what is
being logged yet.
- `./yb_build.sh --cxx-test pg_mini-test --gtest_filter PgMiniTest.IteratorLogPkOnly`
- `./yb_build.sh --cxx-test pg_mini-test --gtest_filter PgMiniTest.IteratorLogTwoNonPkCols`

Reviewers: sergei, bogdan, neha

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D8436