[WIP] Simplified version of DBIter for regular / non-provisional RocksDB
Summary:
We can create a simplified version of DBIter, not supporting RocksDB-level deletes, merges, or
overwrites of the same value, and also not filtering entries with sequence number above the read
point, because we have hybrid time based filtering.

This provides ~4% perf improvement on a microbenchmark.
TBD: TPCC based testing

Test Plan: Jenkins: skip for now

Reviewers: jenkins-bot

Subscribers: ybase, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D8512