[WIP] [#5383] YSQL, Jepsen: Regression in the long-fork test.
Summary:
Commit 02f58124f7015ef0923514bd791db3e58d0e5e37 introduced a regression, causing the `long-fork` test from our Jepsen test suite to fail.
This test is designed to detect a serialization isolation anomaly that looks like this:
```
  T1: (write x 1)
  T2: (write y 1)
  T3: (read x nil) (read y 1)
  T4: (read x 1) (read y nil)

  T3 implies T2 < T1, but T4 implies T1 < T2.
```
This test in a suite used a non-unique index instead of a primary key.
As a part of this diff, we introduce a somewhat simplified version of this test in Java.

IMPORTANT: Add fix details!

---

Resolves #5383

Test Plan: ybd --java-test 'org.yb.pgsql.TestConsistency#longForkWithIndex'

Reviewers: mihnea, mbautin

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D10515